+++ Sources/web/run.py
@@ -1,0 +1,5 @@
+from RightenWeb import app
+from RightenWeb import debugmode
+
+if __name__=="__main__":
+    app.run(debug=debugmode)
